requirements.txt
Adafruit-Blinka==5.0.1
adafruit-circuitpython-bme280==2.2.3
RPi.GPIO==0.7.0